requirements.txt
# Core ML & preprocessing
torch
faiss-cpu
pandas
numpy
scikit-learn
scipy
joblib

# LangChain & LLM Integration
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv

# Optional: visualization, logging, etc.
tqdm
